requirements.txt
alabaster==0.7.12
attrs==19.3.0
Babel==2.9.1
certifi==2019.11.28
chardet==3.0.4
click==7.0
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
idna==2.8
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==1.2.0
Jinja2==2.11.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==8.0.2
networkx==2.4
numpy==1.17.4
opencv-python==4.2.0.32
packaging==19.2
pandas==0.25.3
Pillow==8.3.2
plotly==4.4.1
pluggy==0.13.1
py==1.10.0
Pygments==2.7.4
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
scipy==1.3.3
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7 
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.26.5
wcwidth==0.1.7
zipp==0.6.0
